+++ game/marble/server/scripts/game.cs
@@ -47,6 +47,9 @@ function execServerScripts()
       
    $Server::ScriptsLoaded = true;
    
+   // central game parameters
+   exec("./gameParams.cs");
+
    // game commands (lobby, ready status, and stuff)
    exec("./gameCommands.cs");
    
@@ -127,6 +130,8 @@ function onMissionLoaded()
    else
       $Game::GemCount = countGems(MissionGroup);
       
+   PartyFramework::onMissionLoaded();
+
    initRandomSpawnPoints();
 
    // JMQ: don't start mission yet, wait for command from Lobby 
@@ -140,6 +145,8 @@ function onMissionLoaded()
 
 function onMissionEnded()
 {
+   PartyFramework::onMissionEnded();
+
    // Called by endMission(), right before the mission is destroyed
    // This part of a normal mission cycling or end.
    endGame();
@@ -1013,6 +1020,8 @@ function GameConnection::onClientEnterGame(%this)
 
    if ($Game::packetLoss || $Game::packetLag)
       %this.setSimulatedNetParams($Game::packetLoss,$Game::packetLag/2);
+
+   PartyFramework::onPlayerJoin(%this);
 }
 
 function GameConnection::onClientJoinGame(%this)
@@ -1724,6 +1733,12 @@ function GameConnection::createPlayer(%this, %spawnPoint)
    echo ("Using " @ %physics @ " Physics");
    %player.setPhysics(%physics);
 
+   // Set initial tilt blend
+   if ($Game::TiltGravityMode)
+      %player.setDirectInputBlend(1.0 - $Game::TiltBlend);
+   else
+      %player.setDirectInputBlend(1.0);
+
    // Player setup...
    %spawnPos = getSpawnPosition(%spawnPoint);
    %player.setPosition(%spawnPos, 0.45);
@@ -1747,6 +1762,9 @@ function GameConnection::createPlayer(%this, %spawnPoint)
    // Give the client control of the player
    %this.player = %player;
    %this.setControlObject(%player);
+
+   // Hook into Party Framework
+   PartyFramework::onPlayerSpawn(%player);
 }
 
 
@@ -2449,7 +2467,14 @@ exec("./banana.cs");
 exec("./tilt_gravity.cs");
 exec("./smb_bumpers.cs");
 exec("./smb_switches.cs");
+exec("./smb_conveyors.cs");
 exec("./smb_warpgates.cs");
 exec("./smb_platforms.cs");
 exec("./smb_seesaws.cs");
 exec("./smb_monkey_target.cs");
+exec("./smb_billiards.cs");
+exec("./smb_golf.cs");
+exec("./smb_bowling.cs");
+exec("./party_framework.cs");
+exec("./boss_framework.cs");
+exec("./smb_boss_ape.cs");