+++ game/marble/server/scripts/tilt_gravity.cs
@@ -3,15 +3,36 @@
 //-----------------------------------------------------------------------------
 
 // Toggle for gravity tilt mode
-$Game::TiltGravityMode = true;
+// Values are now defined in gameParams.cs ($Game::UseWorldTilt, $Game::TiltBlend)
 
 function serverCmdToggleTiltGravity(%client)
 {
-   $Game::TiltGravityMode = !$Game::TiltGravityMode;
-   if ($Game::TiltGravityMode)
+   $Game::UseWorldTilt = !$Game::UseWorldTilt;
+   if ($Game::UseWorldTilt)
+   {
       messageClient(%client, 'MsgSystem', '\c0Tilt Gravity Mode: ENABLED');
+      if (isObject(%client.player))
+         %client.player.setDirectInputBlend(1.0 - $Game::TiltBlend);
+   }
    else
+   {
       messageClient(%client, 'MsgSystem', '\c0Tilt Gravity Mode: DISABLED');
+      if (isObject(%client.player))
+      {
+         %client.player.setDirectInputBlend(1.0);
+         // Reset gravity
+         %client.player.setGravityDir("0 0 -19.62", true);
+      }
+   }
+}
+
+function serverCmdSetTiltBlend(%client, %blend)
+{
+   $Game::TiltBlend = %blend;
+   if ($Game::UseWorldTilt && isObject(%client.player))
+   {
+      %client.player.setDirectInputBlend(1.0 - $Game::TiltBlend);
+   }
 }
 
 // Hook into marble update or player update to tilt the gravity vector
@@ -20,15 +41,19 @@ function serverCmdToggleTiltGravity(%client)
 
 function updateTiltGravity(%client, %moveX, %moveY)
 {
-   if (!$Game::TiltGravityMode)
+   if (!$Game::UseWorldTilt)
       return;
+
+   // Also set the blend to the player so they move less with input when tilt is higher
+   if (isObject(%client.player))
+      %client.player.setDirectInputBlend(1.0 - $Game::TiltBlend);
 
    // Base gravity
    %baseZ = -19.62;
 
    // Calculate tilt based on input (moveX/moveY are typically -1 to 1)
    // We want a max tilt of maybe 20 degrees
-   %maxTilt = 0.3; // Multiplier for gravity vector
+   %maxTilt = 0.3 * $Game::TiltBlend; // Multiplier for gravity vector
 
    %gravX = %moveX * %baseZ * %maxTilt;
    %gravY = %moveY * %baseZ * %maxTilt;
@@ -46,7 +71,7 @@ function updateTiltGravity(%client, %moveX, %moveY)
 function tiltGravityLoop()
 {
    cancel($TiltGravitySchedule);
-   if ($Game::TiltGravityMode)
+   if ($Game::UseWorldTilt)
    {
       for (%i = 0; %i < ClientGroup.getCount(); %i++)
       {